+++ src/pages/AddDailyMeal/PurchasingPlan/Shopping/index.tsx
@@ -1,5 +1,5 @@
 import { useParams } from "react-router-dom";
-import ShoppingList from "../AddDailyMeal/ShoppingList";
+import ShoppingList from "../../ShoppingList";
 import useGetPurchasePlan from "./hooks/useGetPurchasePlan";
 import useGetUser from "./hooks/useGetUser";
